requirements.txt
flask==3.0.0
numpy==1.24.3
ffmpeg-python==0.2.0
pyannote.audio==3.1.1
flask-sqlalchemy==3.1.1
scipy==1.11.4
speechbrain==1.0.0
webrtcvad==2.0.10
pydub==0.25.1
redis==5.0.1
celery==5.3.4
python-multipart==0.0.6
requests==2.31.0
